Refactor the debugger code to make more readable
Created a top level MauiDebuggerSession to isolate the heavily maui specific parts
