+++ src/mobile-debug/Debugger/Events/InitializedEvent.cs
@@ -1,0 +1,15 @@
+﻿/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+using VSCodeDebug.Protocol;
+
+namespace VSCodeDebug.Debugger;
+
+// ---- Events -------------------------------------------------------------------------
+
+public class InitializedEvent : Event
+{
+    public InitializedEvent()
+        : base("initialized") { }
+}
