+++ src/mobile-debug/DotnetRunner.cs
@@ -25,7 +25,7 @@ namespace VSCodeDebug
 		public static string DotNetExecutablePath { get; private set; }
 		static string sdkRoot;
 		static string DotNetExeName
-			=> Util.IsWindows ? "dotnet.exe" : "dotnet";
+			=> Utilities.IsWindows ? "dotnet.exe" : "dotnet";
 		
 		public DotnetRunner(string args, string workingDir, CancellationToken cancellationToken, Action<string> outputHandler)
 			: base(DotNetExecutablePath, args, cancellationToken, workingDir, outputHandler)
