+++ src/mobile-debug/ShellProcessRunner.cs
@@ -34,8 +34,8 @@ namespace VSCodeDebug
 			process = new Process();
 			// process.StartInfo.FileName = Util.IsWindows ? "cmd.exe" : (File.Exists("/bin/zsh") ? "/bin/zsh" : "/bin/bash");
 			// process.StartInfo.Arguments = Util.IsWindows ? $"/c \"{executable} {args}\"" : $"-c \"{executable} {args}\"";
-			process.StartInfo.FileName = Util.IsWindows ? executable : (File.Exists("/bin/zsh") ? "/bin/zsh" : "/bin/bash");
-			process.StartInfo.Arguments = Util.IsWindows ? args : $"-c \"{executable} {args}\"";
+			process.StartInfo.FileName = Utilities.IsWindows ? executable : (File.Exists("/bin/zsh") ? "/bin/zsh" : "/bin/bash");
+			process.StartInfo.Arguments = Utilities.IsWindows ? args : $"-c \"{executable} {args}\"";
 
 			if (environmentVariables != null && environmentVariables.Any())
 			{
