--- src/mobile-debug/Util.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.IO;
-
-namespace VsCodeMobileUtil
-{
-	public static class Util
-	{
-		public static bool IsWindows => Environment.OSVersion.Platform == PlatformID.Win32NT;
-
-        //public static void LogToFile(string message)
-        //{
-        //	var desktop = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.DesktopDirectory), "mobile-debug.txt");
-
-        //	File.AppendAllLines(desktop, new[] { message });
-        //}
-
-        public static string FixPathSeparators(string path)
-        {
-            if (Util.IsWindows)
-                return path;
-            return path.Replace("\\", "/");
-        }
-    }
-}
