+++ src/mobile-debug/VSCodeKeys.cs
@@ -1,21 +1,20 @@
 ﻿using System;
-namespace VsCodeMobileUtil {
-	public static class VSCodeKeys {
-		public static class LaunchConfig {
-			public const string Configuration = "projectConfiguration";
-			public const string Output = "projectOutputPath";
-			public const string Platform = "projectPlatform";
-			public const string ProjectPath = "projectPath";
-            public const string ProjectProperties = "projectProperties";
-            public const string ProjectType = "projectType";
-			public const string ProjectTargetFramework = "projectTargetFramework";
-			public const string ProjectIsCore = "projectIsCore";
-			public const string RuntimeIdentifier = "runtimeIdentifier";
-			public const string DeviceId = "device";
-			public const string DebugPort = "debugPort";
-			public const string WorkspaceDirectory = "workspaceUri";
-			public const string AdbEmulatorName = "adbEmulatorName";
+namespace VsCodeMobileUtil; 
+public static class VSCodeKeys {
+	public static class LaunchConfig {
+		public const string Configuration = "projectConfiguration";
+		public const string Output = "projectOutputPath";
+		public const string Platform = "projectPlatform";
+		public const string ProjectPath = "projectPath";
+		public const string ProjectProperties = "projectProperties";
+		public const string ProjectType = "projectType";
+		public const string ProjectTargetFramework = "projectTargetFramework";
+		public const string ProjectIsCore = "projectIsCore";
+		public const string RuntimeIdentifier = "runtimeIdentifier";
+		public const string DeviceId = "device";
+		public const string DebugPort = "debugPort";
+		public const string WorkspaceDirectory = "workspaceUri";
+		public const string AdbEmulatorName = "adbEmulatorName";
 
-		}
 	}
 }
