+++ src/mobile-debug/XCode.cs
@@ -8,186 +8,185 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
-namespace VsCodeMobileUtil
+namespace VsCodeMobileUtil;
+
+public class XCode
 {
-	public class XCode
+	public static List<DeviceData> GetDevices(params string[] targetPlatformIdentifiers)
 	{
-		public static List<DeviceData> GetDevices(params string[] targetPlatformIdentifiers)
-		{
-			var xcode = GetBestXcode();
+		var xcode = GetBestXcode();
 
-			var xcdevice = new FileInfo(Path.Combine(xcode, "Contents/Developer/usr/bin/xcdevice"));
+		var xcdevice = new FileInfo(Path.Combine(xcode, "Contents/Developer/usr/bin/xcdevice"));
 
-			if (!xcdevice.Exists)
-				throw new FileNotFoundException(xcdevice.FullName);
+		if (!xcdevice.Exists)
+			throw new FileNotFoundException(xcdevice.FullName);
 
-			var ir = ProcessRunner.Run(xcdevice,
-				new ProcessArgumentBuilder()
-					.Append("list"));
+		var ir = ProcessRunner.Run(xcdevice,
+			new ProcessArgumentBuilder()
+				.Append("list"));
 
-			var json = string.Join(Environment.NewLine, ir.StandardOutput);
+		var json = string.Join(Environment.NewLine, ir.StandardOutput);
 
-			
-			var xcdevices = JsonConvert.DeserializeObject<List<XcDevice>>(json);
+		
+		var xcdevices = JsonConvert.DeserializeObject<List<XcDevice>>(json);
 
-			var tpidevices = xcdevices
-				.Where(d => (targetPlatformIdentifiers == null || targetPlatformIdentifiers.Length <= 0)
-					|| (targetPlatformIdentifiers.Intersect(d.DotNetPlatforms)?.Any() ?? false));
+		var tpidevices = xcdevices
+			.Where(d => (targetPlatformIdentifiers == null || targetPlatformIdentifiers.Length <= 0)
+				|| (targetPlatformIdentifiers.Intersect(d.DotNetPlatforms)?.Any() ?? false));
 
-			var filteredDevices = tpidevices
-				.Select(d => new DeviceData
-				{
-					IsEmulator = d.Simulator,
-					IsRunning = false,
-					Name = d.Name,
-					Details = d.ModelName + " (" + d.Architecture + ")",
-					Platforms = d.DotNetPlatforms,
-					Serial = d.Identifier,
-					Version = d.OperatingSystemVersion
-				});
-
-			return filteredDevices.ToList();
-		}
-
-		internal static string GetBestXcode()
-		{
-			var selected = GetSelectedXCodePath();
-
-			if (!string.IsNullOrEmpty(selected))
-				return selected;
-
-			return FindXCodeInstalls()?.FirstOrDefault();
-		}
-
-		static string GetSelectedXCodePath()
-		{
-			var r = ProcessRunner.Run(new FileInfo("/usr/bin/xcode-select"), new ProcessArgumentBuilder().Append("-p"));
-
-			var xcodeSelectedPath = string.Join(Environment.NewLine, r.StandardOutput)?.Trim();
-
-			if (!string.IsNullOrEmpty(xcodeSelectedPath))
+		var filteredDevices = tpidevices
+			.Select(d => new DeviceData
 			{
-				var infoPlist = Path.Combine(xcodeSelectedPath, "..", "Info.plist");
-				if (File.Exists(infoPlist))
-				{
-					var info = GetXcodeInfo(
-						Path.GetFullPath(
-							Path.Combine(xcodeSelectedPath, "..", "..")), true);
+				IsEmulator = d.Simulator,
+				IsRunning = false,
+				Name = d.Name,
+				Details = d.ModelName + " (" + d.Architecture + ")",
+				Platforms = d.DotNetPlatforms,
+				Serial = d.Identifier,
+				Version = d.OperatingSystemVersion
+			});
 
-					if (info != null)
-						return info?.Path;
-				}
-			}
+		return filteredDevices.ToList();
+	}
 
-			return null;
-		}
+	internal static string GetBestXcode()
+	{
+		var selected = GetSelectedXCodePath();
 
-		static readonly string[] LikelyPaths = new[]
+		if (!string.IsNullOrEmpty(selected))
+			return selected;
+
+		return FindXCodeInstalls()?.FirstOrDefault();
+	}
+
+	static string GetSelectedXCodePath()
+	{
+		var r = ProcessRunner.Run(new FileInfo("/usr/bin/xcode-select"), new ProcessArgumentBuilder().Append("-p"));
+
+		var xcodeSelectedPath = string.Join(Environment.NewLine, r.StandardOutput)?.Trim();
+
+		if (!string.IsNullOrEmpty(xcodeSelectedPath))
 		{
-			"/Applications/Xcode.app",
-			"/Applications/Xcode-beta.app",
-		};
-
-		static IEnumerable<string> FindXCodeInstalls()
-		{
-			foreach (var p in LikelyPaths)
+			var infoPlist = Path.Combine(xcodeSelectedPath, "..", "Info.plist");
+			if (File.Exists(infoPlist))
 			{
-				var i = GetXcodeInfo(p, false)?.Path;
-				if (i != null)
-					yield return i;
+				var info = GetXcodeInfo(
+					Path.GetFullPath(
+						Path.Combine(xcodeSelectedPath, "..", "..")), true);
+
+				if (info != null)
+					return info?.Path;
 			}
 		}
 
-		static (string Path, bool Selected)? GetXcodeInfo(string path, bool selected)
-		{
-			var versionPlist = Path.Combine(path, "Contents", "version.plist");
+		return null;
+	}
 
-			if (File.Exists(versionPlist))
+	static readonly string[] LikelyPaths = new[]
+	{
+		"/Applications/Xcode.app",
+		"/Applications/Xcode-beta.app",
+	};
+
+	static IEnumerable<string> FindXCodeInstalls()
+	{
+		foreach (var p in LikelyPaths)
+		{
+			var i = GetXcodeInfo(p, false)?.Path;
+			if (i != null)
+				yield return i;
+		}
+	}
+
+	static (string Path, bool Selected)? GetXcodeInfo(string path, bool selected)
+	{
+		var versionPlist = Path.Combine(path, "Contents", "version.plist");
+
+		if (File.Exists(versionPlist))
+		{
+			return (path, selected);
+		}
+		else
+		{
+			var infoPlist = Path.Combine(path, "Contents", "Info.plist");
+
+			if (File.Exists(infoPlist))
 			{
 				return (path, selected);
 			}
-			else
-			{
-				var infoPlist = Path.Combine(path, "Contents", "Info.plist");
-
-				if (File.Exists(infoPlist))
-				{
-					return (path, selected);
-				}
-			}
-			return null;
 		}
+		return null;
 	}
+}
 
-	public class XcDevice
-	{
+public class XcDevice
+{
 
-		public const string PlatformMacOsx = "com.apple.platform.macosx";
-		public const string PlatformiPhoneSimulator = "com.apple.platform.iphonesimulator";
-		public const string PlatformAppleTvSimulator = "com.apple.platform.appletvsimulator";
-		public const string PlatformAppleTv = "com.apple.platform.appletvos";
-		public const string PlatformWatchSimulator = "com.apple.platform.watchsimulator";
-		public const string PlatformiPhone = "com.apple.platform.iphoneos";
-		public const string PlatformWatch = "com.apple.platform.watchos";
+	public const string PlatformMacOsx = "com.apple.platform.macosx";
+	public const string PlatformiPhoneSimulator = "com.apple.platform.iphonesimulator";
+	public const string PlatformAppleTvSimulator = "com.apple.platform.appletvsimulator";
+	public const string PlatformAppleTv = "com.apple.platform.appletvos";
+	public const string PlatformWatchSimulator = "com.apple.platform.watchsimulator";
+	public const string PlatformiPhone = "com.apple.platform.iphoneos";
+	public const string PlatformWatch = "com.apple.platform.watchos";
 
-		[JsonProperty("simulator")]
-		public bool Simulator { get; set; }
+	[JsonProperty("simulator")]
+	public bool Simulator { get; set; }
 
-		[JsonProperty("operatingSystemVersion")]
-		public string OperatingSystemVersion { get; set; }
+	[JsonProperty("operatingSystemVersion")]
+	public string OperatingSystemVersion { get; set; }
 
-		[JsonProperty("available")]
-		public bool Available { get; set; }
+	[JsonProperty("available")]
+	public bool Available { get; set; }
 
-		[JsonProperty("platform")]
-		public string Platform { get; set; }
+	[JsonProperty("platform")]
+	public string Platform { get; set; }
 
-		public bool IsiOS
-			=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformiPhone) || Platform.Equals(PlatformiPhoneSimulator));
-		public bool IsTvOS
-			=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformAppleTv) || Platform.Equals(PlatformAppleTvSimulator));
-		public bool IsWatchOS
-			=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformWatch) || Platform.Equals(PlatformWatchSimulator));
-		public bool IsOsx
-			=> !string.IsNullOrEmpty(Platform) && Platform.Equals(PlatformMacOsx);
+	public bool IsiOS
+		=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformiPhone) || Platform.Equals(PlatformiPhoneSimulator));
+	public bool IsTvOS
+		=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformAppleTv) || Platform.Equals(PlatformAppleTvSimulator));
+	public bool IsWatchOS
+		=> !string.IsNullOrEmpty(Platform) && (Platform.Equals(PlatformWatch) || Platform.Equals(PlatformWatchSimulator));
+	public bool IsOsx
+		=> !string.IsNullOrEmpty(Platform) && Platform.Equals(PlatformMacOsx);
 
-		public string[] DotNetPlatforms
-			=> Platform switch {
-				PlatformiPhone => new[] { "ios" },
-				PlatformiPhoneSimulator => new[] { "ios" },
-				PlatformAppleTv => new[] { "tvos" },
-				PlatformAppleTvSimulator => new[] { "tvos" },
-				PlatformWatch => new [] { "watchos" },
-				PlatformWatchSimulator => new[] { "watchos" },
-				PlatformMacOsx => new[] {"macos", "maccatalyst"},
-				_ => new string[0]
-			};
+	public string[] DotNetPlatforms
+		=> Platform switch {
+			PlatformiPhone => new[] { "ios" },
+			PlatformiPhoneSimulator => new[] { "ios" },
+			PlatformAppleTv => new[] { "tvos" },
+			PlatformAppleTvSimulator => new[] { "tvos" },
+			PlatformWatch => new [] { "watchos" },
+			PlatformWatchSimulator => new[] { "watchos" },
+			PlatformMacOsx => new[] {"macos", "maccatalyst"},
+			_ => new string[0]
+		};
 
-		[JsonProperty("modelCode")]
-		public string ModelCode { get; set; }
+	[JsonProperty("modelCode")]
+	public string ModelCode { get; set; }
 
-		[JsonProperty("identifier")]
-		public string Identifier { get; set; }
+	[JsonProperty("identifier")]
+	public string Identifier { get; set; }
 
-		[JsonProperty("architecture")]
-		public string Architecture { get; set; }
+	[JsonProperty("architecture")]
+	public string Architecture { get; set; }
 
-		public string RuntimeIdentifier
-			=> Architecture switch
-			{
-				_ => "iossimulator-x64"
-			};
+	public string RuntimeIdentifier
+		=> Architecture switch
+		{
+			_ => "iossimulator-x64"
+		};
 
-		[JsonProperty("modelUTI")]
-		public string ModelUTI { get; set; }
+	[JsonProperty("modelUTI")]
+	public string ModelUTI { get; set; }
 
-		[JsonProperty("modelName")]
-		public string ModelName { get; set; }
+	[JsonProperty("modelName")]
+	public string ModelName { get; set; }
 
-		[JsonProperty("name")]
-		public string Name { get; set; }
+	[JsonProperty("name")]
+	public string Name { get; set; }
 
-		[JsonProperty("interface")]
-		public string Interface { get; set; }
-	}
+	[JsonProperty("interface")]
+	public string Interface { get; set; }
 }
